% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{calculate_kurtosis_skewness}
\alias{calculate_kurtosis_skewness}
\title{Calcular Kurtosis y Skewness con Gráficos}
\usage{
calculate_kurtosis_skewness(data)
}
\arguments{
\item{data}{Un vector numérico o un data frame que contiene los datos a analizar.}
}
\value{
Un data frame que contiene la kurtosis, skewness y si los datos se desvían
        de una distribución normal para cada variable.
}
\description{
Esta función calcula la kurtosis y skewness de un vector de datos numéricos
y genera histogramas y boxplots para visualización. También determina si los
datos se desvían de una distribución normal basándose en umbrales predefinidos.
}
\examples{
data <-  c(2.3, 1.5, 3.7, 2.8, 4.2, 5.6)

calculate_kurtosis_skewness(data)
}
