% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{calculate_kurtosis_skewness_all}
\alias{calculate_kurtosis_skewness_all}
\title{Calcular Kurtosis y Skewness para Todas las Variables Numéricas}
\usage{
calculate_kurtosis_skewness_all(data)
}
\arguments{
\item{data}{Un data frame que contiene las variables a analizar.}
}
\value{
Un data frame que contiene la kurtosis, skewness y si los datos se desvían
        de una distribución normal para cada variable numérica.
}
\description{
Esta función toma un data frame y calcula la kurtosis y skewness de todas las variables
numéricas en él. Además, genera histogramas y boxplots para visualizar las distribuciones
de las variables numéricas y determina si los datos se desvían de una distribución normal.
}
\examples{
data <- data.frame(
  var1 = c(2.3, 1.5, 3.7, 2.8, 4.2, 5.6),
  var2 = c(7.8, 6.4, 8.2, 7.1, 9.5, 10.2),
  var3 = c(12.1, 14.8, 15.6, 13.2, 16.7, 11.9)
)
calculate_kurtosis_skewness_all(data)
}
