% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{consistencia}
\alias{consistencia}
\title{Obtener características de variables en un data frame}
\usage{
consistencia(data)
}
\arguments{
\item{data}{Un data frame. El conjunto de datos del cual se obtendrán las características de las variables.}
}
\value{
Un nuevo data frame que contiene las características de las variables en el conjunto de datos.
}
\description{
Esta función calcula diversas características de las variables en un data frame y las devuelve en un nuevo data frame.
}
\examples{
# Supongamos que tenemos el siguiente data frame
datos <- data.frame(
  edad = c(25, 30, 22),
  altura = c(1.75, 1.68, 1.80),
  ciudad = c("Madrid", "Paris", NA)
)

# Llamamos a la función para obtener las características de las variables
caracteristicas_df <- consistencia(datos)

# El nuevo data frame contendrá las características de las variables
#   variable    tipo  media  desviación_estandar  minimo  máximo  valores_únicos  longitud  valores_faltantes  porcentaje_valores_faltantes
#1     edad numeric 25.667            4.163332      22      30               3         3                  0                           0\%
#2   altura numeric  1.776            0.057735    1.68    1.80               3         3                  0                           0\%
#3   ciudad  factor     NA                 NA    <NA>    <NA>               3         3                  1                     33.333\%

}
