% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{convertir_nombres_mayusculas}
\alias{convertir_nombres_mayusculas}
\title{Convertir nombres de columnas a mayúsculas}
\usage{
convertir_nombres_mayusculas(data)
}
\arguments{
\item{data}{Un data frame. El conjunto de datos cuyas columnas se convertirán a mayúsculas.}
}
\value{
Un nuevo data frame con los mismos datos que el original, pero con los nombres de las columnas en mayúsculas.
}
\description{
Esta función toma un data frame y convierte los nombres de las columnas a letras mayúsculas.
}
\examples{
# Supongamos que tenemos el siguiente data frame
#   nombre  edad  ciudad
#0  Alice    25  Madrid
#1    Bob    30   Paris
#2  Carol    22  Berlin

# Llamamos a la función para convertir los nombres de las columnas a mayúsculas
nuevo_df <- convertir_nombres_mayusculas(datos)

# El nuevo data frame tendrá los nombres de las columnas en mayúsculas
#   NOMBRE  EDAD  CIUDAD
#0  Alice    25  Madrid
#1    Bob    30   Paris
#2  Carol    22  Berlin

}
