% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{detectar_valores_atipicos}
\alias{detectar_valores_atipicos}
\title{Detectar Valores Atípicos en Variables Numéricas}
\usage{
detectar_valores_atipicos(data)
}
\arguments{
\item{data}{Un data frame que contiene las variables numéricas a analizar.}
}
\value{
Un data frame que muestra las variables, los índices de las filas con valores atípicos y los valores atípicos detectados.
}
\description{
Esta función permite detectar valores atípicos en variables numéricas de un conjunto de datos.
}
\details{
La función calcula los límites usando el rango intercuartílico (IQR) y detecta valores atípicos que caen fuera de estos límites.
         También genera un boxplot para cada variable numérica con los valores atípicos resaltados en rojo.
}
\examples{
data_ejemplo <- data.frame(
  edad = c(25, 30, 35, 40, 45, 50, 55, 60, 65, 70),
  ingreso = c(500, 1000, 1200, 1500, 2000, 2200, 2300, 2500, 2800, 3000)
)
detectar_valores_atipicos(data_ejemplo)

}
