% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{detectar_valores_ausentes}
\alias{detectar_valores_ausentes}
\title{Detectar valores ausentes en un conjunto de datos}
\usage{
detectar_valores_ausentes(data)
}
\arguments{
\item{data}{Un data frame o matriz que contiene los datos a analizar.}
}
\value{
Un data frame con las siguientes columnas:
\describe{
  \item{Variable}{Nombre de la variable en la que se detectaron valores ausentes.}
  \item{Cantidad_Ausentes}{Cantidad total de valores ausentes en la variable.}
  \item{Porcentaje_Ausentes}{Porcentaje de valores ausentes en la variable (redondeado a 2 decimales).}
}
}
\description{
Esta función detecta y reporta la cantidad y el porcentaje de valores ausentes en cada
variable de un conjunto de datos.
}
\examples{
# Crear un data frame de ejemplo
data <- data.frame(
  Col1 = c(1, NA, 3, 4),
  Col2 = c(NA, 2, 3, NA),
  Col3 = c(NA, NA, NA, NA)
)

# Aplicar la función para detectar valores ausentes
resultados <- detectar_valores_ausentes(data)

}
