% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{estandarizar_variables}
\alias{estandarizar_variables}
\title{Estandarizar variables numéricas en una base de datos}
\usage{
estandarizar_variables(data)
}
\arguments{
\item{data}{Un objeto de tipo data.frame o tibble que contiene los datos a procesar.}
}
\value{
Un nuevo objeto de tipo data.frame o tibble con las mismas dimensiones que el
  objeto de entrada, pero con las variables numéricas estandarizadas. Las columnas numéricas
  originales son reemplazadas por las versiones estandarizadas y renombradas agregando el
  sufijo "_estandarizada" al nombre original.
}
\description{
Esta función toma como entrada una base de datos y estandariza las variables numéricas
presentes en ella. La estandarización consiste en transformar cada variable numérica
para que tenga media cero y desviación estándar igual a uno. Esta transformación es útil
cuando se desean comparar variables que tienen diferentes escalas, ya que coloca todas
las variables en una misma escala relativa.
}
\examples{
data_ejemplo <- data.frame(
  edad = c(25, 30, 22),
  ingreso = c(50000, 60000, 48000),
  gasto = c(2000, 2500, 1800)
)

data_ejemplo_estandarizado <- estandarizar_variables(data_ejemplo)
str(data_ejemplo_estandarizado)

}
