% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{factores_a_string}
\alias{factores_a_string}
\title{Convertir factores de una base de datos a cadenas de texto}
\usage{
factores_a_string(data)
}
\arguments{
\item{data}{Un objeto de tipo data.frame o tibble que contiene los datos a procesar.}
}
\value{
Un nuevo objeto de tipo data.frame o tibble con las mismas dimensiones que el
  objeto de entrada, pero con las columnas que eran factores convertidas a cadenas de texto.
  El objeto devuelto tendrá las mismas columnas y nombres que el objeto de entrada.
}
\description{
Esta función toma como entrada una base de datos y convierte todas las columnas que
sean de tipo factor a cadenas de texto (character). Los factores son un tipo especial
de datos en R que representan variables categóricas con un conjunto finito de niveles
posibles. La conversión a cadenas de texto facilita su manipulación y análisis posterior,
ya que muchas operaciones y gráficas en R trabajan mejor con datos en formato de caracteres.
}
\examples{
data_ejemplo <- data.frame(
  edad = c(25, 30, 22),
  genero = factor(c("masculino", "femenino", "masculino")),
  ciudad = factor(c("New York", "Los Angeles", "Chicago"))
)

data_ejemplo_convertido <- factores_a_string(data_ejemplo)
str(data_ejemplo_convertido)

}
