% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{mi_funcion}
\alias{mi_funcion}
\title{Mi función}
\usage{
mi_funcion(x)
}
\arguments{
\item{x}{Un vector numérico.}
}
\value{
El resultado de la función.
}
\description{
Esta función realiza una tarea específica.
}
\examples{
mi_funcion(1:10)
}
