% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{reemplazar_valores}
\alias{reemplazar_valores}
\title{Reemplazar Valores en una Columna de una Base de Datos}
\usage{
reemplazar_valores(data, columna, valor_original, valor_nuevo)
}
\arguments{
\item{data}{Un data frame que representa la base de datos en la que se realizará el reemplazo.}

\item{columna}{El nombre de la columna en la que se reemplazarán los valores.}

\item{valor_original}{El valor que se desea reemplazar en la columna.}

\item{valor_nuevo}{El nuevo valor con el que se reemplazará el valor original.}
}
\value{
Un data frame con los valores reemplazados en la columna especificada.
}
\description{
Esta función permite reemplazar valores en una columna específica de una base de datos.
}
\details{
La función busca el valor original en la columna especificada y lo reemplaza por el valor nuevo.
}
\examples{
# Crear un data frame de ejemplo
data <- data.frame(
  Nombre = c("Ana", "Juan", "María", "Pedro"),
  Edad = c(25, 30, 22, 30)
)

# Reemplazar el valor 30 en la columna "Edad" por 31
nuevo_data <- reemplazar_valores(data, "Edad", 30, 31)

}
