% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{resumen}
\alias{resumen}
\title{Resumen de características de variables en un conjunto de datos}
\usage{
resumen(data)
}
\arguments{
\item{data}{Un dataframe o matriz de datos.}
}
\value{
Un dataframe con el resumen de características de cada variable.
}
\description{
Esta función calcula varias características descriptivas para cada variable en el conjunto de datos.
}
\examples{
data <- data.frame(
  Var1 = c(1, 2, 3, NA, 5),
  Var2 = c("A", "B", "A", "C", "B"),
  Var3 = c(10.2, 9.5, 11.1, 10.8, NA)
)
result <- resumen(data)
print(result)

}
\keyword{descriptivas}
\keyword{estadísticas}
\keyword{resumen}
\keyword{variables}
