% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{test_lilliefors}
\alias{test_lilliefors}
\title{Prueba de normalidad  de Lilliefors}
\usage{
test_lilliefors(data, alpha = 0.05)
}
\arguments{
\item{data}{A numeric vector.}

\item{alpha}{The significance level for the test (default is 0.05).}
}
\value{
A list with the test statistic, p-value, and normality assessment.
}
\description{
Performs the Lilliefors test (Kolmogorov-Smirnov test for normality) on a numeric vector.
}
\examples{
data <- c(1.2, 2.5, 3.1, 4.3, 5.6, 6.8, 7.2, 8.4, 9.7, 10.9)
lilliefors_result <- test_lilliefors(data)
print(lilliefors_result)

}
\seealso{
\code{\link{nortest::lillie.test}}
}
\author{
[OSCAR GONZALEZ FRUTOS]
}
