% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{test_lilliefors_all}
\alias{test_lilliefors_all}
\title{Prueba de normalidad de Lilliefors para variables numéricas}
\usage{
test_lilliefors_all(data, alpha = 0.05)
}
\arguments{
\item{data}{Un dataframe o matriz que contiene las variables a ser probadas.}

\item{alpha}{Nivel de significancia para la prueba. El valor predeterminado es 0.05.}
}
\value{
Un dataframe con los resultados de la prueba de normalidad para cada columna numérica.
La columna 'Estadistica' contiene los valores de la estadística de prueba,
la columna 'Valor_p' contiene los valores p asociados,
y la columna 'Normalidad' indica si la variable se considera normal o no.
}
\description{
Esta función realiza la prueba de normalidad de Lilliefors en cada columna numérica del conjunto de datos proporcionado.
}
\examples{
data <- data.frame(
  Var1 = rnorm(100),
  Var2 = rnorm(100),
  Var3 = rpois(100, lambda = 3)
)
result <- test_lilliefors_all(data)
print(result)

}
\seealso{
\link{nortest::lillie.test}
}
\keyword{estadísticas}
\keyword{normalidad}
\keyword{prueba}
